+++ mlxtend/feature_selection/tests/test_sequential_feature_selector.py
@@ -500,14 +500,68 @@ def test_regression_in_range():
 
 
 def test_clone_params_fail():
+
+    class Perceptron(object):
+
+        def __init__(self, eta=0.1, epochs=50, random_seed=None,
+                     print_progress=0):
+
+            self.eta = eta
+            self.epochs = epochs
+            self.random_seed = random_seed
+            self.print_progress = print_progress
+            self._is_fitted = False
+
+        def _fit(self, X, y, init_params=True):
+            self._check_target_array(y, allowed={(0, 1)})
+            y_data = np.where(y == 0, -1., 1.)
+
+            if init_params:
+                self.b_, self.w_ = self._init_params(
+                    weights_shape=(X.shape[1], 1),
+                    bias_shape=(1,),
+                    random_seed=self.random_seed)
+                self.cost_ = []
+
+            rgen = np.random.RandomState(self.random_seed)
+            for i in range(self.epochs):
+                errors = 0
+
+                for idx in self._yield_minibatches_idx(
+                        rgen=rgen,
+                        n_batches=y_data.shape[0], data_ary=y_data, shuffle=True):
+
+                    update = self.eta * (y_data[idx] -
+                                         self._to_classlabels(X[idx]))
+                    self.w_ += (update * X[idx]).reshape(self.w_.shape)
+                    self.b_ += update
+                    errors += int(update != 0.0)
+
+                if self.print_progress:
+                    self._print_progress(iteration=i + 1,
+                                         n_iter=self.epochs,
+                                         cost=errors)
+                self.cost_.append(errors)
+            return self
+
+        def _net_input(self, X):
+            """ Net input function """
+            return (np.dot(X, self.w_) + self.b_).flatten()
+
+        def _to_classlabels(self, X):
+            return np.where(self._net_input(X) < 0.0, -1., 1.)
+
+        def _predict(self, X):
+            return np.where(self._net_input(X) < 0.0, 0, 1)
+
     if sys.version_info >= (3, 0):
         objtype = 'class'
     else:
         objtype = 'type'
 
     expect = ("Cannot clone object"
-              " '<class 'mlxtend.classifier."
-              "softmax_regression.SoftmaxRegression'>'"
+              " '<class 'test_sequential_feature_selector."
+              "test_clone_params_fail.<locals>.Perceptron'>'"
               " (type <%s 'type'>): it does not seem to be a"
               " scikit-learn estimator as it does not"
               " implement a 'get_params' methods.") % objtype
@@ -515,7 +569,7 @@ def test_clone_params_fail():
     assert_raises(TypeError,
                   expect,
                   SFS,
-                  SoftmaxRegression,
+                  Perceptron,
                   scoring='accuracy',
                   k_features=3,
                   clone_estimator=True)
@@ -532,7 +586,7 @@ def test_clone_params_pass():
                floating=False,
                scoring='accuracy',
                cv=0,
-               clone_estimator=False,
+               clone_estimator=True,
                verbose=0,
                n_jobs=1)
     sfs1 = sfs1.fit(X, y)